Identify outermost functions
This identifies the top-level functions defined in the contract based on
their function signature hash.  The start of each function is clearly
delinated within the code.  However, it may not be the case that blocks
are grouped according to the function they are in.

Furthermore, have removed `{:verify false}` from a number of (mostly
small) blocks.  These are ones which Dafny can verify already.  Others
(especially those containing `assert` statements) remain problematic.
Whenever an `assert` statement is contained within a method, it usually
means some appropriate precondition must be propagated through.