Verified allowance()!
This verifies the `allowance()` function.  In the end, I discovered a
fix for the various Dafny problems I was having with `block_0_0x03d9`.
The issue it appears is the use of `And()` within a block.  In the
DafnyEVM this turns into operations over a bitvector (`bv256`) which are
known to be expensive.

Therefore, have added custom `And()` implementations (`AndU160`,
`AndU8`) which are designed for the particular uses found in this
contract.  Namely, masking a value to implement a coercion (e.g. going
from a `u256` to a `u160` by masking against `MAX_U160`).