Working through necessary preconditions
This has been a fairly painful process working backwards through the
proof.  Things to do:

(1) `block_0_0x0508` appears to be where the mysterious loop bound is
assignment to memory location `0x60`, and the frame pointer is
increased.

(2) There looks to be some kind of bug in the Dafny Proof generator as
we have a loop involving `block_0_0x0566` which is malformed.  Looks
like the target address is calculated incorrectly.

That's all folks!